requirements.txt
pyparsing
appdirs
setuptools==40.1.0
cryptography==2.8
bcrypt==3.1.7
PyNaCl==1.3.0
Fabric3==1.14.post1

